Use UUID rather than name when storing Version data (#1119)
* Use UUID rather than name when storing Version data

Deleting and recreating a Version can result in COS data being lost.

The working theory is that when the COS object was deleted, COS internally marks it as deleted and cleans it up at a later time. When the cleanup occurs, COS checks to see if the object is still marked as deleted and then fully removes it.
Separately, the object is recreated. If the recreation occurs between when the cleanup checks if it's still deleted and fully removes it, the end result is that the object is removed even though both the deletion and creation reported success.

I.e. a very small window for data loss, but possible, and has been seen once.

https://github.ibm.com/alchemy-containers/satellite-config/issues/1504

* dep fix (minimist)